Test Class Structure Restructured: Reorganization and restructuring of the test class structure related to the institution. XML Test Suite Files Updated: Modification made in the XML Test Suite files to include specific test cases. Updated 'Extract User ID' Method: Update made in the method responsible for extracting user IDs to align with recent API changes.
