+++ src/test/java/payload/Helper.java
@@ -11,30 +11,7 @@ import java.util.Map;
 import java.util.Random;
 
 public class Helper {
-
-    public static Assignment createRandomAssignment() {
-        int MIN_LATITUDE = 35;
-        int MAX_LATITUDE = 42;
-        int MIN_LONGITUDE = 26;
-        int MAX_LONGITUDE = 45;
-
-        Faker faker = new Faker();
-        Assignment assignment = new Assignment();
-        assignment.setFirstName(faker.name().firstName());
-        assignment.setLastName(faker.name().lastName());
-
-        PhoneNumber phoneNumber = new PhoneNumber();
-        phoneNumber.setLineNumber(generateLineNumber());
-        phoneNumber.setCountryCode("90");
-        assignment.setPhoneNumber(phoneNumber);
-
-        assignment.setDescription(faker.commerce().productName());
-        assignment.setLatitude(faker.number().randomDouble(6, MIN_LATITUDE, MAX_LATITUDE));
-        assignment.setLongitude(faker.number().randomDouble(6, MIN_LONGITUDE, MAX_LONGITUDE));
-
-        return assignment;
-    }
-    public static User getUser(String userID){
+    public static User getUser(String userID) {
         Response response = InstitutionEndpoints.getUser(userID);
         if (response.getStatusCode() == 200) {
             if (response.jsonPath().get("response") != null) {
@@ -51,19 +28,8 @@ public class Helper {
     }
 
     public static UserCredentials createNewUser() {
-        Faker faker;
-        User userPayload;
-        PhoneNumber phoneNumber;
-        faker = new Faker();
-        userPayload = new User();
-        phoneNumber = new PhoneNumber();
-        userPayload.setFirstName(faker.name().firstName());
-        userPayload.setLastName(faker.name().lastName());
-        phoneNumber.setLineNumber(generateLineNumber());
-        phoneNumber.setCountryCode("90");
-        userPayload.setPhoneNumber(phoneNumber);
-        Response response = InstitutionEndpoints.createAUser(userPayload);
-
+        User user = createUserPayload();
+        Response response = InstitutionEndpoints.createAUser(user);
         if (response.getStatusCode() == 200) {
             return response.then()
                     .extract().jsonPath().getObject("response", UserCredentials.class);
@@ -71,9 +37,9 @@ public class Helper {
             throw new RuntimeException("User creation failed with status code: " + response.getStatusCode());
         }
     }
+
     public static UserCredentials createNewUser(User userPayload) {
         Response response = InstitutionEndpoints.createAUser(userPayload);
-
         if (response.getStatusCode() == 200) {
             return response.then()
                     .extract().jsonPath().getObject("response", UserCredentials.class);
@@ -82,20 +48,15 @@ public class Helper {
         }
     }
 
-    public static User createUserPayload(){
+    public static User createUserPayload() {
         Faker faker;
         User userPayload;
-        PhoneNumber phoneNumber;
         faker = new Faker();
         userPayload = new User();
-        phoneNumber = new PhoneNumber();
         userPayload.setFirstName(faker.name().firstName());
         userPayload.setLastName(faker.name().lastName());
-        phoneNumber.setLineNumber(generateLineNumber());
-        phoneNumber.setCountryCode("90");
-        userPayload.setPhoneNumber(phoneNumber);
+        userPayload.setPhoneNumber(createPhoneNumber());
         return userPayload;
-
     }
 
     public static Map<String, String> createLoginPayload(String username, String password) {
@@ -105,7 +66,7 @@ public class Helper {
         return payload;
     }
 
-    public static RequestBodyAssignments createRequestBodyAssignments(String page,String pageSize) {
+    public static RequestBodyAssignments createRequestBodyAssignments(int page, int pageSize) {
         RequestBodyAssignments requestBodyAssignments = new RequestBodyAssignments();
         Pagination pagination = new Pagination();
         pagination.setPage(page);
@@ -113,9 +74,10 @@ public class Helper {
         requestBodyAssignments.setPagination(pagination);
         return requestBodyAssignments;
     }
-    public static Assignment createANewAssignment(){
-        Faker faker=new Faker();
-        Assignment assignment=new Assignment();
+
+    public static Assignment createANewAssignment() {
+        Faker faker = new Faker();
+        Assignment assignment = new Assignment();
         assignment.setFirstName(faker.name().firstName());
         assignment.setLastName(faker.name().lastName());
         assignment.setPhoneNumber(createPhoneNumber());
@@ -125,8 +87,9 @@ public class Helper {
         InstitutionEndpoints.createAnAssignment(assignment);
         return assignment;
     }
-    public static PhoneNumber createPhoneNumber(){
-        PhoneNumber phoneNumber=new PhoneNumber();
+
+    public static PhoneNumber createPhoneNumber() {
+        PhoneNumber phoneNumber = new PhoneNumber();
         phoneNumber.setLineNumber(generateLineNumber());
         phoneNumber.setCountryCode("90");
         return phoneNumber;
@@ -166,53 +129,96 @@ public class Helper {
         return id;
     }
 
+    public static String extractUserIdByDescriptionAndName(String description, String firstName, String lastName) {
+        int currentPage = 1;
+        int totalPageCount = Integer.MAX_VALUE;
+
+        while (currentPage <= totalPageCount) {
+            Response assignmentResponse = InstitutionEndpoints.listAssignments(createRequestBodyAssignments(currentPage, 10));
+            totalPageCount = assignmentResponse.jsonPath().getInt("response.totalPageCount");
+
+            String assignmentId = findAssignmentIdInContent(assignmentResponse, description, firstName, lastName);
+            if (assignmentId != null) {
+                return assignmentId;
+            }
+
+            currentPage++;
+            if (currentPage > totalPageCount) {
+                break;
+            }
+        }
+        return null;
+    }
+
+    private static String findAssignmentIdInContent(Response response, String description, String firstName, String lastName) {
+        List<Map<String, Object>> content = response.jsonPath().getList("response.content");
+
+        for (Map<String, Object> entry : content) {
+            String desc = (String) entry.get("description");
+            String fName = (String) entry.get("firstName");
+            String lName = (String) entry.get("lastName");
+
+            if (desc != null && desc.equals(description) && fName != null && fName.equals(firstName) && lName != null && lName.equals(lastName)) {
+                return (String) entry.get("id");
+            }
+        }
+        return null;
+    }
+
+
     public static String getUserID(String username) {
-        String pagination = "{\"pagination\":{\"page\":1,\"pageSize\":10}}";
         int currentPage = 1;
 
         while (true) {
-            Response response = InstitutionEndpoints.listUsers(pagination);
+            Pagination pagination = setPagination(currentPage, 10);
+            RequestBodyAssignments requestBodyAssignments = new RequestBodyAssignments();
+            requestBodyAssignments.setPagination(pagination);
+            Response response = InstitutionEndpoints.listUsers(requestBodyAssignments);
+            response.then().log().body();
             String userID = extractUserIdFromTheUserList(response, username);
-
             if (userID != null) {
                 return userID;
             }
-
             int currentPageNumber = response.jsonPath().getInt("response.pageNumber");
             int totalPageCount = response.jsonPath().getInt("response.totalPageCount");
-
             if (currentPageNumber >= totalPageCount) {
                 break;
             }
-            pagination = "{\"pagination\":{\"page\":" + (currentPage + 1) + ",\"pageSize\":10}}";
             currentPage++;
         }
 
         return null;
     }
-    public static String getUserIDByFirstName(String firstname) {
-        String pagination = "{\"pagination\":{\"page\":1,\"pageSize\":10}}";
+
+    public static String getUserIDByFirstName(String firstName) {
+        RequestBodyAssignments requestBodyAssignments = new RequestBodyAssignments();
+        Pagination pagination = setPagination(1, 10);
+        requestBodyAssignments.setPagination(setPagination(1, 10));
         int currentPage = 1;
-
         while (true) {
-            Response response = InstitutionEndpoints.listUsers(pagination);
-            String userID = extractUserIdFromTheUserListByFirstname(response, firstname);
-
+            Response response = InstitutionEndpoints.listUsers(requestBodyAssignments);
+            String userID = extractUserIdFromTheUserListByFirstname(response, firstName);
             if (userID != null) {
                 return userID;
             }
-
             int currentPageNumber = response.jsonPath().getInt("response.pageNumber");
             int totalPageCount = response.jsonPath().getInt("response.totalPageCount");
-
             if (currentPageNumber >= totalPageCount) {
                 break;
             }
-            pagination = "{\"pagination\":{\"page\":" + (currentPage + 1) + ",\"pageSize\":10}}";
+            pagination.setPage(currentPage + 1);
+            requestBodyAssignments.setPagination(pagination);
             currentPage++;
         }
 
         return null;
+    }
+
+    public static Pagination setPagination(int page, int pageSize) {
+        Pagination pagination = new Pagination();
+        pagination.setPageSize(pageSize);
+        pagination.setPage(page);
+        return pagination;
     }
 
     private static String extractUserIdFromTheUserList(Response response, String username) {
@@ -225,6 +231,7 @@ public class Helper {
         }
         return null;
     }
+
     private static String extractUserIdFromTheUserListByFirstname(Response response, String firstname) {
         List<Map<String, Object>> userList = response.jsonPath().getList("response.content");
         for (Map<String, Object> user : userList) {
@@ -245,6 +252,7 @@ public class Helper {
         location.setLongitude(longitude);
         return location;
     }
+
     public static Location generateLocationTR() {
         Random rand = new Random();
         double latitude = 38 + (40 - 38) * rand.nextDouble();
@@ -265,10 +273,12 @@ public class Helper {
                 "    \"supportStatus\": \"" + supportStatus + "\"\n" +
                 "}";
     }
+
     private static double generateRandomCoordinate(int min, int max) {
         Random random = new Random();
         return min + (max - min) * random.nextDouble();
     }
+
     public static String generateString(int length) {
         String characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*()_+-=[]{}|;:,.<>?/";
         Random random = new Random();
