+++ src/test/java/runners/InstitutionTestSuite.xml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="All Test Suite">
+    <test verbose="2" preserve-order="true" name="Institution Test Suite">
+        <classes>
+            <class name="tests.auth.InstitutionAuthServiceTest">
+                <methods>
+                    <include name="getTokenForValidAdmin"/>
+                    <include name="getTokenForInvalidAdmin"/>
+                    <include name="adminTokenRefresh"/>
+                    <include name="adminInvalidateToken"/>
+                </methods>
+            </class>
+            <class name="tests.institution.adminmanagementservice.PostAdminsTest">
+                <methods>
+                    <include name="listAdminsWithValidPageAndPageSize"/>
+                    <include name="listAdminsWithEmptyPageValue"/>
+                    <include name="listAdminsWithEmptyPageSizeValue"/>
+                    <include name="listAdminsWithMissingPageField"/>
+                    <include name="listAdminsWithMissingPageSizeField"/>
+                    <include name="listAdminsWithMissingPageAndPageSizeField"/>
+                    <include name="listAdminsWithPageSizeValueNot10"/>
+                </methods>
+            </class>
+            <class name="tests.institution.assignmentmanagementservice.DeleteAssignmentTest">
+                <methods>
+                    <include name="deleteAssignment"/>
+                    <include name="deleteAssignmentNegative"/>
+                </methods>
+            </class>
+            <class name="tests.institution.assignmentmanagementservice.GetAssignmentTest">
+                <methods>
+                    <include name="getAssignmentPositive"/>
+                    <include name="getAssignmentAfterDelete"/>
+                </methods>
+            </class>
+            <class name="tests.institution.assignmentmanagementservice.PostAssignmentTest">
+                <methods>
+                    <include name="createAnAssignmentPositive"/>
+                    <include name="createAssignmentWithInvalidCountryCode"/>
+                    <include name="createAssignmentWithInvalidPhoneNumber"/>
+                    <include name="createAssignmentWithInvalidLongitude"/>
+                    <include name="createAssignmentWithInvalidLatitude"/>
+                    <include name="createAssignmentWithMissingFirstName"/>
+                    <include name="createAssignmentWithMissingLastName"/>
+                    <include name="createAssignmentWithMissingDescription"/>
+                    <include name="createAssignmentWithMissingPhoneNumber"/>
+                    <include name="createAssignmentWithMissingLatitude"/>
+                    <include name="createAssignmentWithMissingLongitude"/>
+                    <include name="createAssignmentWithInvalidDataCombination"/>
+                </methods>
+            </class>
+            <class name="tests.institution.assignmentmanagementservice.PostAssignmentsTest">
+                <methods>
+                    <include name="listAssignmentsWithFilters"/>
+                    <include name="postAssignmentsPagination"/>
+                    <include name="postAssignmentsMissingPagination"/>
+                </methods>
+            </class>
+            <class name="tests.institution.assignmentmanagementservice.PutAssignmentTest">
+                <methods>
+                    <include name="updateAssignmentFirstname"/>
+                </methods>
+            </class>
+            <class name="tests.institution.usermanagementservice.DeleteUserTest">
+                <methods>
+                    <include name="deleteUser"/>
+                    <include name="deleteUserNegative"/>
+                </methods>
+            </class>
+            <class name="tests.institution.usermanagementservice.GetUserTest">
+                <methods>
+                    <include name="getUser"/>
+                </methods>
+            </class>
+            <class name="tests.institution.usermanagementservice.PostUserTest">
+                <methods>
+                    <include name="createAUser"/>
+                    <include name="createUserNegative"/>
+                </methods>
+            </class>
+            <class name="tests.institution.usermanagementservice.PostUsersTest">
+                <methods>
+                    <include name="listUsersPositive"/>
+                </methods>
+            </class>
+            <class name="tests.institution.usermanagementservice.PutUserTest">
+                <methods>
+                    <include name="updateUserAsPassive"/>
+                    <include name="updateUserAsActive"/>
+                    <include name="updateUserWithInvalidRole"/>
+                    <include name="updateUserWithBlankRole"/>
+                    <include name="updateUserWithBlankStatus"/>
+                </methods>
+            </class>
+        </classes>
+    </test>
+</suite>
