--- src/test/java/runners/TestSuite.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="All Test Suite">
-    <listeners>
-        <listener class-name="utility.ExtentReportManager"></listener>
-    </listeners>
-    <test verbose="2" preserve-order="true"
-          name="AYS API Test Suite">
-        <classes>
-            <class name="tests.institution.usermanagementservice.PostUserTest"></class>
-            <class name="tests.auth.InstitutionAuthServiceTest"></class>
-            <class name="tests.auth.UserAuthServiceTest"></class>
-            <class name="tests.institution.adminmanagementservice.InstitutionManagementTest"></class>
-            <class name="tests.institution.assignmentmanagementservice.PostAssignmentsServiceTest"></class>
-            <class name="tests.institution.assignmentmanagementservice.PostAssignmentTest"></class>
-            <class name="tests.institution.assignmentmanagementservice.GetUpdateDeleteAssignmentTest"></class>
-            <class name="tests.institution.usermanagementservice.DeleteUserTest"></class>
-            <class name="tests.institution.usermanagementservice.GetUserTest"></class>
-            <class name="tests.institution.usermanagementservice.PostUserTest"></class>
-            <class name="tests.institution.usermanagementservice.PostUsersTest"></class>
-            <class name="tests.institution.usermanagementservice.PutUserTest"></class>
-
-        </classes>
-
-
-    </test>
-</suite>
