+++ src/test/java/tests/institution/assignmentmanagementservice/DeleteAssignmentTest.java
@@ -1,4 +1,52 @@
 package tests.institution.assignmentmanagementservice;
 
+import endpoints.InstitutionEndpoints;
+import io.restassured.response.Response;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.testng.Assert;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
+import payload.Assignment;
+import payload.Helper;
+
+import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.equalTo;
+
 public class DeleteAssignmentTest {
+    Assignment assignment=new Assignment();
+    String assignmentId;
+    Logger logger = LogManager.getLogger(this.getClass());
+    @BeforeMethod
+    public void setup(){
+        assignment= Helper.createANewAssignment();
+        assignmentId = Helper.extractUserIdByDescriptionAndName(assignment.getDescription(),assignment.getFirstName(),assignment.getLastName());
+    }
+    @Test()
+    public void deleteAssignment() {
+        Response response = InstitutionEndpoints.deleteAssignment(assignmentId);
+        response.then()
+                .statusCode(200)
+                .contentType("application/json")
+                .body("httpStatus", equalTo("OK"))
+                .body("isSuccess", equalTo(true));
+        Assert.assertTrue(InstitutionEndpoints.getAssignment(assignmentId).jsonPath().getString("message").contains("ASSIGNMENT NOT EXIST!"));
+    }
+
+
+    @Test()
+    public void deleteAssignmentNegative() {
+        logger.info("IAMS_46 is running");
+        InstitutionEndpoints.deleteAssignment(assignmentId);
+        Response response = InstitutionEndpoints.deleteAssignment(assignmentId);
+        response.then()
+                .statusCode(404)
+                .contentType("application/json")
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("NOT_FOUND"))
+                .body("header", equalTo("NOT EXIST"))
+                .body("message", containsString("ASSIGNMENT NOT EXIST!"))
+                .body("isSuccess", equalTo(false));
+    }
+
 }
