+++ src/test/java/tests/user/assignmentmanagementservice/GetAssignmentSummaryTest.java
@@ -1,4 +1,4 @@
-package tests.user.userassignmentmanagementservice;
+package tests.user.assignmentmanagementservice;
 
 import endpoints.UserEndpoints;
 import io.restassured.response.Response;
