+++ src/test/java/tests/user/locationmanagementservice/PostUserLocationTest.java
@@ -1,10 +1,10 @@
-package tests.user.userlocationmanagementservice;
+package tests.user.locationmanagementservice;
 
 import endpoints.UserEndpoints;
 import io.restassured.response.Response;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.testng.annotations.BeforeClass;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import payload.Assignment;
 import payload.Helper;
@@ -12,7 +12,6 @@ import payload.Location;
 import payload.UserCredentials;
 
 import static org.hamcrest.Matchers.*;
-import static org.hamcrest.Matchers.equalTo;
 
 public class PostUserLocationTest {
     UserCredentials userCredentials;
@@ -21,17 +20,18 @@ public class PostUserLocationTest {
     Assignment assignment;
 
 
-    @BeforeClass
+    @BeforeMethod
     public void setup() {
         userCredentials = Helper.createNewUser();
         location = new Location();
-        assignment=Helper.createANewAssignment();
+        assignment = Helper.createANewAssignment();
 
     }
+
     @Test()
     public void updateLocationWithReservedAssignment() {
         logger.info("Test case UMS_05 is running");
-        location = Helper.generateLocation(38, 40, 28, 43);
+        location = Helper.generateLocationTR();
         Helper.setSupportStatus("READY", userCredentials.getUsername(), userCredentials.getPassword());
         UserEndpoints.searchAssignment(location, userCredentials.getUsername(), userCredentials.getPassword());
         Response response = UserEndpoints.updateLocation(location, userCredentials.getUsername(), userCredentials.getPassword());
@@ -43,10 +43,14 @@ public class PostUserLocationTest {
                 .body("message", containsString("USER LOCATION CANNOT BE UPDATED BECAUSE USER DOES NOT HAVE AN ASSIGNMENT IN PROGRESS! "))
                 .body("isSuccess", equalTo(false));
     }
+
     @Test()
     public void updateLocationWithAssignedAssignment() {
         logger.info("Test case UMS_08 is running");
-        location = Helper.generateLocation(38, 40, 28, 43);
+        location = Helper.generateLocationTR();
+        Helper.setSupportStatus("READY", userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.searchAssignment(location, userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.approveAssignment(userCredentials.getUsername(), userCredentials.getPassword());
         Response response = UserEndpoints.updateLocation(location, userCredentials.getUsername(), userCredentials.getPassword());
         response.then()
                 .statusCode(500)
@@ -56,10 +60,15 @@ public class PostUserLocationTest {
                 .body("message", containsString("USER LOCATION CANNOT BE UPDATED BECAUSE USER DOES NOT HAVE AN ASSIGNMENT IN PROGRESS! "))
                 .body("isSuccess", equalTo(false));
     }
+
     @Test()
     public void updateLocationAfterStart() {
         logger.info("Test case UMS_12 is running");
-        location = Helper.generateLocation(38, 40, 28, 43);
+        location = Helper.generateLocationTR();
+        Helper.setSupportStatus("READY", userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.searchAssignment(location, userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.approveAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.startAssignment(userCredentials.getUsername(), userCredentials.getPassword());
         Response response = UserEndpoints.updateLocation(location, userCredentials.getUsername(), userCredentials.getPassword());
         response.then()
                 .statusCode(200)
